Finished Implementing Sequence Generator
The sequence generator is now fully functional. The user can specify the number of game sequences to be generated (by changing the variable SEQ_GEN_NUM), and these sequences are written to a text file called SEQ_GEN_NUM-sequences.txt.

Additionally, I added some quality-of-life features that enable the user to monitor the progress of sequence generation (A basic indicator that notifies the user via the terminal every 5000 sequences generated that the program is still running), as well as listing the maximum and minimum sequence lengths (Max cannot be greater than 102 moves, Min cannot be less than 2 moves), and the P1/P2 Win percentages (should hover around 50%/50%).